change: orm

+++ src/route_fn.rs
@@ -2,12 +2,12 @@ use axum::{http::StatusCode, Json, extract::Path};
 
 use crate::scrape;
 
-pub async fn novel_statistics(Path(novel_id): Path<u64>) -> (StatusCode, Json<scrape::NovelStatistic>) {
+pub async fn novel_statistics(Path(novel_id): Path<i32>) -> (StatusCode, Json<scrape::NovelStatistic>) {
     let novel = scrape::get_novel_detail(novel_id).await.unwrap();
     (StatusCode::OK, Json(novel))
 }
 
-pub async fn novel_clicks(Path(novel_id): Path<u64>) -> (StatusCode, Json<(u64, u64)>) {
+pub async fn novel_clicks(Path(novel_id): Path<i32>) -> (StatusCode, Json<(i32, i32)>) {
     let data = scrape::get_chapter_clicks(novel_id).await.unwrap();
     (StatusCode::OK, Json(data))
 }
